fix: unify HTTP/HTTPS applications into single configurable app
- Merge app.ts and https-app.ts into unified application
- Add HTTPS_ENABLED environment variable for HTTP/HTTPS switching
- Implement dynamic protocol detection in /health endpoint
- Add conditional security headers, CORS, and session configuration
- Split GitHub Actions workflow into separate HTTP/HTTPS test jobs
- Remove duplicate https-app.ts file
- Update README.md with new unified architecture and configuration
- Update package.json scripts to use environment variables
