+++ src/app.ts
@@ -1,4 +1,6 @@
 import express from 'express';
+import https from 'https';
+import fs from 'fs';
 import session from 'express-session';
 import passport from 'passport';
 import path from 'path';
@@ -32,35 +34,75 @@ const __dirname = path.dirname(__filename);
 
 dotenv.config();
 
+// 環境変数からHTTPS有効/無効を判定
+const HTTPS_ENABLED = process.env.HTTPS_ENABLED === 'true';
+const HTTP_PORT = process.env.HTTP_PORT || 3000;
+const HTTPS_PORT = process.env.HTTPS_PORT || 3443;
+const PORT = HTTPS_ENABLED ? HTTPS_PORT : HTTP_PORT;
+
 const app = express();
-const PORT = process.env.PORT || 3000;
 
 // MCP サーバーインスタンス
 let mcpServer: Server | null = null;
 let mcpTransport: StreamableHTTPServerTransport | null = null;
 
-// セキュリティヘッダー
-app.use(helmet({
+// SSL証明書のパス
+const SSL_KEY_PATH = path.join(__dirname, '../ssl/localhost.key');
+const SSL_CERT_PATH = path.join(__dirname, '../ssl/localhost.crt');
+
+// SSL証明書の存在確認
+const checkSSLCertificates = (): boolean => {
+  try {
+    return fs.existsSync(SSL_KEY_PATH) && fs.existsSync(SSL_CERT_PATH);
+  } catch {
+    return false;
+  }
+};
+
+// セキュリティヘッダー（HTTPS/HTTPに応じて設定）
+const helmetConfig: any = {
   contentSecurityPolicy: {
     directives: {
       defaultSrc: ["'self'"],
       styleSrc: ["'self'", "'unsafe-inline'"],
       scriptSrc: ["'self'"],
-      connectSrc: ["'self'", "https:"],
+      connectSrc: ["'self'", "https:", "wss:"],
       imgSrc: ["'self'", "data:", "https:"],
     },
   },
   crossOriginEmbedderPolicy: false
-}));
+};
 
-// CORS設定
+// HTTPS専用のセキュリティヘッダー
+if (HTTPS_ENABLED) {
+  helmetConfig.hsts = {
+    maxAge: 31536000,
+    includeSubDomains: true,
+    preload: true
+  };
+}
+
+app.use(helmet(helmetConfig));
+
+// CORS設定（HTTPS/HTTPに応じて設定）
+const corsOrigins = [
+  'https://claude.ai',
+  /^http:\/\/localhost:\d+$/,
+  /^http:\/\/127\.0\.0\.1:\d+$/,
+  /^http:\/\/\[::1\]:\d+$/
+];
+
+// HTTPS有効時はHTTPS URLも許可
+if (HTTPS_ENABLED) {
+  corsOrigins.push(
+    /^https:\/\/localhost:\d+$/,
+    /^https:\/\/127\.0\.0\.1:\d+$/,
+    /^https:\/\/\[::1\]:\d+$/
+  );
+}
+
 app.use(cors({
-  origin: [
-    'https://claude.ai',
-    /^http:\/\/localhost:\d+$/,
-    /^http:\/\/127\.0\.0\.1:\d+$/,
-    /^http:\/\/\[::1\]:\d+$/
-  ],
+  origin: corsOrigins,
   credentials: true,
   methods: ['GET', 'POST', 'OPTIONS'],
   allowedHeaders: ['Content-Type', 'Authorization', 'X-Session-Id']
@@ -70,11 +112,16 @@ app.use(express.json({ limit: '10mb' }));
 app.use(express.urlencoded({ extended: true, limit: '10mb' }));
 app.use(express.static(path.join(__dirname, '../public')));
 
+// セッション設定（HTTPS/HTTPに応じて設定）
 app.use(session({
   secret: process.env.SESSION_SECRET || 'default-secret-key',
   resave: false,
   saveUninitialized: false,
-  cookie: { secure: false }
+  cookie: { 
+    secure: HTTPS_ENABLED, // HTTPS時のみセキュアクッキーを有効
+    httpOnly: true,
+    maxAge: 24 * 60 * 60 * 1000 // 24時間
+  }
 }));
 
 app.use(passport.initialize());
@@ -85,6 +132,17 @@ app.use('/api', ticketRoutes);
 
 app.get('/', (_req, res) => {
   res.sendFile(path.join(__dirname, '../public/index.html'));
+});
+
+// ヘルスチェックエンドポイント（HTTP/HTTPS対応）
+app.get('/health', (req, res) => {
+  const protocol = req.secure || req.headers['x-forwarded-proto'] === 'https' || HTTPS_ENABLED ? 'HTTPS' : 'HTTP';
+  res.json({
+    status: 'OK',
+    protocol: protocol,
+    timestamp: new Date().toISOString(),
+    service: 'Authlete Study Session Ticket Service'
+  });
 });
 
 // MCP サーバー初期化
@@ -231,12 +289,56 @@ const startServer = async (): Promise<void> => {
     await DatabaseConfig.initialize();
     AuthService.initializePassport();
     await initializeMCPServer();
-    
-    app.listen(PORT, () => {
-      console.log(`Server running on http://localhost:${PORT}`);
-      console.log(`MCP endpoint: http://localhost:${PORT}/mcp`);
-      console.log(`MCP health check: http://localhost:${PORT}/mcp/health`);
-    });
+
+    if (HTTPS_ENABLED) {
+      // HTTPS モード
+      
+      // SSL証明書の確認
+      if (!checkSSLCertificates()) {
+        console.error('❌ SSL証明書が見つかりません');
+        console.log('以下のコマンドでSSL証明書を生成してください:');
+        console.log('  npm run generate-ssl');
+        console.log('または:');
+        console.log('  ./scripts/generate-ssl-cert.sh');
+        process.exit(1);
+      }
+
+      // SSL証明書を読み込み
+      const privateKey = fs.readFileSync(SSL_KEY_PATH, 'utf8');
+      const certificate = fs.readFileSync(SSL_CERT_PATH, 'utf8');
+      const credentials = { key: privateKey, cert: certificate };
+
+      // HTTPサーバー（HTTPSへのリダイレクト用）
+      const httpApp = express();
+      httpApp.use((req, res) => {
+        const httpsUrl = `https://${req.headers.host?.replace(/:\d+/, `:${HTTPS_PORT}`)}${req.url}`;
+        res.redirect(301, httpsUrl);
+      });
+
+      // HTTPSサーバーを起動
+      const httpsServer = https.createServer(credentials, app);
+      
+      httpsServer.listen(HTTPS_PORT, () => {
+        console.log(`🔒 HTTPS Server running on https://localhost:${HTTPS_PORT}`);
+        console.log(`🔗 MCP endpoint: https://localhost:${HTTPS_PORT}/mcp`);
+        console.log(`💚 MCP health check: https://localhost:${HTTPS_PORT}/mcp/health`);
+        console.log(`📊 App health check: https://localhost:${HTTPS_PORT}/health`);
+      });
+
+      // HTTPサーバーを起動（リダイレクト用）
+      httpApp.listen(HTTP_PORT, () => {
+        console.log(`↗️  HTTP Server running on http://localhost:${HTTP_PORT} (redirects to HTTPS)`);
+      });
+
+    } else {
+      // HTTP モード
+      app.listen(PORT, () => {
+        console.log(`🌐 HTTP Server running on http://localhost:${PORT}`);
+        console.log(`🔗 MCP endpoint: http://localhost:${PORT}/mcp`);
+        console.log(`💚 MCP health check: http://localhost:${PORT}/mcp/health`);
+        console.log(`📊 App health check: http://localhost:${PORT}/health`);
+      });
+    }
   } catch (error) {
     console.error('Failed to start server:', error);
     process.exit(1);
